requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
deprecation==2.1.0
distro==1.8.0
gotrue==1.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
lxml==4.9.3
mbedmgr @ file:///home/kayce/repos/mbedmgr
openai==1.1.1
packaging==23.2
postgrest==0.13.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.0
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supafunc==0.3.1
tiktoken==0.5.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
websockets==10.4
